Move easing code to Utils namespace
