+++ Scripts/Data/ChartInfo/Storyboardable.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 using UnityEngine.Serialization;
+using JANOARG.Shared.Utils.Animation;
 
 // TODO: REWRITE THIS ASAP
