+++ Scripts/Utils/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2caa40abfa169777dbb68d698fca5f68
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
