+++ Scripts/Utils/Animation/Easing.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d25c2cb80346b0ae59e1db37e7715d17
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
