+++ Scripts/Utils/FastMath.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 146359c54d97f4e4298e80987a174063
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
